\name{Robot0001-package}
\alias{Robot0001-package}
\alias{Robot0001}
\docType{package}
\title{
\packageTitle{Robot0001}
}
\description{
\packageDescription{Robot0001}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Robot0001}
\packageIndices{Robot0001}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Robot0001}

Maintainer: \packageMaintainer{Robot0001}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
